BE Project
Face Expo Reader is the tool that is capable of automatically analyzing
facial                 expressions, providing users with an objective
assessment of a person’s emotion.  A facial expression is one or more
motions or positions of the muscles in the skin. These movements convey
the emotional state of the individual to observers.